Complete risc-v implementation
This took so long after several attempts to try and get loading a 64 bit immediate to work - I just couldn't manage to do it. So I'll leave that for another time.

It doesn't help that there's no documentation on doing so because it's generally frowned upon. It's unfortunate as dasm is in a special place where it'd be very convenient to do so to use as a JIT.